move fixtures to static and add generation script
