+++ .eslintrc.js
@@ -114,6 +114,7 @@ module.exports = {
         'packages/private-build-infra/src/**/*.js',
         'packages/unpublished-test-infra/src/**/*.js',
         'packages/unpublished-eslint-rules/src/**/*.js',
+        'packages/unpublished-relationship-performance-test-app/fixtures/**/*.js',
         'packages/*/.ember-cli.js',
         'packages/*/.eslintrc.js',
         'packages/*/.template-lintrc.js',
