+++ packages/unpublished-relationship-performance-test-app/app/adapters/application.js
@@ -2,11 +2,9 @@ import EmberObject from '@ember/object';
 
 import { resolve } from 'rsvp';
 
-import CARS_PAYLOAD from '../utils/generate-fixtures-for-materialization-scenario';
-
 export default class ApplicationMockAdapter extends EmberObject {
   findAll() {
-    return resolve(CARS_PAYLOAD);
+    return fetch('./fixtures/materialization.json').then((response) => response.json());
   }
   shouldReloadAll() {
     return false;
