+++ packages/unpublished-relationship-performance-test-app/app/routes/destroy.js
@@ -3,12 +3,11 @@ import { run } from '@ember/runloop';
 
 import { all } from 'rsvp';
 
-import { createParentPayload } from '../utils/create-parent-payload';
-
 export default Route.extend({
-  model() {
+  async model() {
     performance.mark('start-data-generation');
-    const payload = createParentPayload(2000, 500);
+    const payload = await fetch('./fixtures/destroy.json').then((r) => r.json());
+
     performance.mark('start-push-payload');
     const result = this.store.push(payload);
     performance.mark('start-destroy-records');
