+++ packages/unpublished-relationship-performance-test-app/app/templates/application.hbs
@@ -1,1 +1,7 @@
 <h1>Performance Test Application</h1>
+<ol>
+  <li><LinkTo @route='materialization'>Materialization</LinkTo></li>
+  <li><LinkTo @route='unload'>Unload</LinkTo></li>
+  <li><LinkTo @route='destroy'>Destroy</LinkTo></li>
+  <li><LinkTo @route='add-children'>Add Children</LinkTo></li>
+</ol>
